requirements.txt
fastapi
supabase
google-generativeai
uvicorn
